backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
asyncpg>=0.30.0
alembic>=1.14.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.12
Pillow>=11.0.0
aiofiles>=24.1.0
websockets>=14.0
psycopg2-binary>=2.9.10
